# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ACE-Step Suno-Compatible API
  version: "1.5.0"
  description: |
    100% Suno-API-compatible REST interface powered by ACE-Step 1.5.

    All music-generation endpoints are **asynchronous**: the initial POST returns a
    `taskId` that you poll with the corresponding `record-info` GET until `status`
    reaches `SUCCESS` or `FAILED`.

    Output audio files are served under `/audio/<filename>` as static files.

    ### Authentication
    When the server is configured with `API_KEY`, every request **must** include:
    ```
    Authorization: Bearer <api_key>
    ```
    If `API_KEY` is unset, the header is optional and all requests are accepted.

  contact:
    name: ACE-Step Project
    url: https://github.com/ace-step/ACE-Step-1.5
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:7080
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: >
        Pass your API key as a Bearer token.
        Leave header absent when the server has no API_KEY configured.

  # ── Shared enums ────────────────────────────────────────────────────────────
  schemas:
...
    TaskStatus:
      type: string
      description: |
        Lifecycle of an async task:
        - `PENDING` – queued, not yet started
        - `GENERATING` – actively processing
        - `TEXT_SUCCESS` – LLM phase completed, DiT phase pending
        - `FIRST_SUCCESS` – first audio sample ready (batch mode)
        - `SUCCESS` – all outputs ready
        - `FAILED` – terminal error
        - `CREATE_TASK_FAILED` – task could not be enqueued
      enum:
        - PENDING
        - GENERATING
        - TEXT_SUCCESS
        - FIRST_SUCCESS
        - SUCCESS
        - FAILED
        - CREATE_TASK_FAILED
    TaskType:
      type: string
      enum: [GENERATE, LYRICS, VOCAL_REMOVAL, WAV, MP4, COVER, SOUNDS]

    # ── Common response envelope ───────────────────────────────────────────────
    ApiResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: HTTP-like status code (200 = success).
          example: 200
        msg:
          type: string
          example: success
        data:
          description: Payload; null on error.
          nullable: true
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            code:
              example: 400
            msg:
              example: Bad Request
            data:
              nullable: true
              example: null
    TaskCreatedData:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
          description: Opaque identifier – use it with `record-info` endpoints.
          example: a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c

    # ── Audio output entry (sunoData item) ────────────────────────────────────
    SunoAudioEntry:
      type: object
      properties:
        id:
          type: string
          description: Unique audio identifier (UUID hex).
          example: d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9
        audioUrl:
          type: string
          nullable: true
          description: Relative URL to stream / download the audio file.
          example: /audio/a3f1b24c_0.flac
        streamAudioUrl:
          type: string
          nullable: true
          example: /audio/a3f1b24c_0.flac
        imageUrl:
          type: string
          nullable: true
          description: Cover art URL (null unless cover was generated).
        prompt:
          type: string
          description: Lyrics or caption embedded in the output.
        modelName:
          type: string
          example: ACE-Step-v1.5
        title:
          type: string
          example: Track 1
        tags:
          type: string
          description: Style tags used for generation.
        createTime:
          type: string
          description: UTC timestamp of generation.
          example: "2026-04-11 12:00:00"
        duration:
          type: number
          nullable: true
          description: Duration of the audio in seconds.
          example: 30.4

    # ── Task status response ───────────────────────────────────────────────────
    TaskDetailData:
      type: object
      required: [taskId, status, type]
      properties:
        taskId:
          type: string
        param:
          type: string
          description: JSON-serialised original request parameters.
        response:
          nullable: true
          description: Generation result payload (sunoData array) once SUCCESS.
          type: object
          properties:
            taskId:
              type: string
            sunoData:
              type: array
              items:
                $ref: "#/components/schemas/SunoAudioEntry"
        status:
          $ref: "#/components/schemas/TaskStatus"
        type:
          $ref: "#/components/schemas/TaskType"
        errorCode:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true

    # ── Request bodies ─────────────────────────────────────────────────────────
...
    GenerateCoverRequest:
      type: object
      required: [taskId, callBackUrl]
      properties:
        taskId:
          type: string
          description: Task ID of the generation job to generate cover art for.
        callBackUrl:
          type: string
          format: uri
        prompt:
          type: string
          maxLength: 300
          nullable: true
          description: >
            Optional explicit image generation prompt.
            When omitted, the style tags from the original generation task are used.
        style:
          type: string
          maxLength: 100
          nullable: true
...
    CoverImageEntryData:
      type: object
      properties:
        id:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          example: complete
        errorMessage:
          type: string

  # ── Reusable responses ──────────────────────────────────────────────────────
  responses:
    TaskCreated:
      description: Task accepted; poll `record-info` with the returned `taskId`.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TaskCreatedData"
          example:
            code: 200
            msg: success
            data:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
...
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            msg: "customMode=true requires: style, title"
            data: null
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            msg: "Invalid API key."
            data: null
    NotFound:
      description: Task or resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 404
            msg: "Task not found"
            data: null
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/suno/cover/generate:
    post:
      summary: Generate album cover art
      operationId: generate_cover
      tags: [Cover Art]
      description: |
        Generates a cover image for a completed generation task.
        If `prompt` is omitted, the style tags from the original task are used.

        Image generation priority:
        1. Stable Diffusion (if `diffusers` is installed and GPU available)
        2. PIL gradient placeholder
        3. 1×1 black transparent PNG stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCoverRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              callBackUrl: "https://example.com/webhook"
              prompt: "A neon city skyline at dusk, atmospheric, cinematic"
              style: "Synthwave"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/suno/cover/record-info:
    get:
      summary: Poll cover art task
      operationId: get_cover_details
      tags: [Cover Art]
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cover art task status.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/TaskDetailData"
                          - type: object
                            properties:
                              response:
                                type: object
                                properties:
                                  sunoData:
                                    type: array
                                    items:
                                      $ref: "#/components/schemas/CoverImageEntryData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Style ───────────────────────────────────────────────────────────────────
...